If applied: - this separates out the controllers in editClientDetailsController - removes editClientDetailsController - updates the existing tests - removes editClientDetailsController unit spec
