+++ routes/editClientDetails.ts
@@ -1,36 +1,30 @@
 import express from 'express';
-import {
-  getEditClientName,
-  postEditClientName,
-  getEditClientEmailAddress,
-  postEditClientEmailAddress,
-  getEditClientPhoneNumber,
-  postEditClientPhoneNumber
-} from '#src/scripts/controllers/editClientDetailsController.js';
+import { getEditClientName, postEditClientName } from '#src/scripts/controllers/editClientNameController.js'
+import { getEditClientDateOfBirth, postEditClientDateOfBirth } from '#src/scripts/controllers/editClientDateOfBirthController.js';
+import { getEditClientPhoneNumber,postEditClientPhoneNumber} from '#src/scripts/controllers/editClientPhoneNumberController.js';
+import { getEditClientEmailAddress, postEditClientEmailAddress, } from '#src/scripts/controllers/editClientEmailAddressController.js'
 import { getEditClientAddress, postEditClientAddress } from '#src/scripts/controllers/editClientAddressController.js';
-import { 
-  getEditClientDateOfBirth, 
-  postEditClientDateOfBirth 
-} from '#src/scripts/controllers/editClientDateOfBirthController.js';
+
+import { validateEditClientDateOfBirth } from '#src/middlewares/clientDateOfBirthSchema.js';
 import { validateEditClientPhoneNumber } from '#src/middlewares/phoneNumberSchema.js';
 import { validateEditClientAddress } from '#src/middlewares/clientAddressSchema.js';
-import { validateEditClientDateOfBirth } from '#src/middlewares/clientDateOfBirthSchema.js';
 
 const router = express.Router();
 
 router.get('/:caseReference/client-details/change/name', getEditClientName);
 router.post('/:caseReference/client-details/change/name', postEditClientName);
 
-router.get('/:caseReference/client-details/change/email-address', getEditClientEmailAddress);
-router.post('/:caseReference/client-details/change/email-address', postEditClientEmailAddress);
+router.get('/:caseReference/client-details/change/date-of-birth', getEditClientDateOfBirth);
+router.post('/:caseReference/client-details/change/date-of-birth', validateEditClientDateOfBirth(), postEditClientDateOfBirth);
 
 router.get('/:caseReference/client-details/change/phone-number', getEditClientPhoneNumber);
 router.post('/:caseReference/client-details/change/phone-number', validateEditClientPhoneNumber(), postEditClientPhoneNumber);
 
+router.get('/:caseReference/client-details/change/email-address', getEditClientEmailAddress);
+router.post('/:caseReference/client-details/change/email-address', postEditClientEmailAddress);
+
 router.get('/:caseReference/client-details/change/address', getEditClientAddress);
 router.post('/:caseReference/client-details/change/address', validateEditClientAddress(), postEditClientAddress);
 
-router.get('/:caseReference/client-details/change/date-of-birth', getEditClientDateOfBirth);
-router.post('/:caseReference/client-details/change/date-of-birth', validateEditClientDateOfBirth(), postEditClientDateOfBirth);
 
 export default router;
