+++ tests/unit/controllers/editClientNameController.spec.ts
@@ -1,0 +1,138 @@
+/**
+ * Edit Client Name Controller Tests
+ * 
+ * Tests the Express.js controllers for client details editing functionality.
+ * Covers HTTP request/response handling for name and email editing forms including:
+ * - GET route handlers for form display
+ * - POST route handlers for form submission  
+ * - API integration and error handling
+ * - CSRF token management
+ * - Form data processing and validation integration
+ * 
+ * Testing Level: Unit (Controller Layer)
+ * Component: Express.js Controllers
+ * Dependencies: apiService, form validation helpers
+ */
+
+import { describe, it, beforeEach, afterEach } from 'mocha';
+import { expect } from 'chai';
+import * as sinon from 'sinon';
+import type { Request, Response, NextFunction } from 'express';
+import {
+  getEditClientName,
+  postEditClientName
+} from '#src/scripts/controllers/editClientNameController.js';
+import { apiService } from '#src/services/apiService.js';
+// Import to get global type declarations for axiosMiddleware
+import '#utils/axiosSetup.js';
+
+// Define the RequestWithMiddleware interface for testing
+interface RequestWithMiddleware extends Request {
+  axiosMiddleware: any;
+  csrfToken?: () => string;
+}
+
+describe('Edit Client Name Controller', () => {
+  let req: Partial<RequestWithMiddleware>;
+  let res: any;
+  let next: any;
+  let renderStub: sinon.SinonStub;
+  let redirectStub: sinon.SinonStub;
+  let statusStub: sinon.SinonStub;
+  let apiServiceGetStub: sinon.SinonStub;
+  let apiServiceUpdateStub: sinon.SinonStub;
+
+  beforeEach(() => {
+    req = {
+      params: { caseReference: 'TEST123' },
+      body: {},
+      axiosMiddleware: {} as any,
+      csrfToken: () => 'test-csrf-token'
+    } as Partial<RequestWithMiddleware>;
+
+    renderStub = sinon.stub();
+    redirectStub = sinon.stub();
+    statusStub = sinon.stub().returns({ render: renderStub });
+
+    res = {
+      render: renderStub,
+      redirect: redirectStub,
+      status: statusStub
+    };
+
+    next = sinon.stub();
+
+    // Stub the API service methods
+    apiServiceGetStub = sinon.stub(apiService, 'getClientDetails');
+    apiServiceUpdateStub = sinon.stub(apiService, 'updateClientDetails');
+  });
+
+  afterEach(() => {
+    sinon.restore();
+  });
+
+  describe('getEditClientName', () => {
+    it('should render name editing form with pre-populated client data and CSRF protection', async () => {
+      // Arrange
+      const mockApiResponse = {
+        status: 'success',
+        data: {
+          fullName: 'John Doe',
+          caseReference: 'TEST123'
+        }
+      };
+
+      apiServiceGetStub.resolves(mockApiResponse);
+
+      // Act
+      await getEditClientName(req as RequestWithMiddleware, res as Response, next);
+
+      // Assert
+      expect(apiServiceGetStub.calledOnce).to.be.true;
+      expect(renderStub.calledWith('case_details/edit-client-name.njk')).to.be.true;
+    });
+
+    it('should delegate API errors to Express error handling middleware', async () => {
+      // Arrange
+      const error = new Error('API Error');
+      apiServiceGetStub.rejects(error);
+
+      // Act
+      await getEditClientName(req as RequestWithMiddleware, res as Response, next);
+
+      // Assert
+      expect(next.calledOnce).to.be.true;
+    });
+  });
+
+  describe('postEditClientName', () => {
+    it('should process successful client name update and redirect to case details', async () => {
+      // Arrange
+      req.body = { fullName: 'Jane Smith' };
+
+      apiServiceUpdateStub.resolves({
+        status: 'success',
+        data: { fullName: 'Jane Smith' }
+      });
+
+      // Act
+      await postEditClientName(req as RequestWithMiddleware, res as Response, next);
+
+      // Assert
+      expect(apiServiceUpdateStub.calledOnce).to.be.true;
+      expect(redirectStub.calledWith('/cases/TEST123/client-details')).to.be.true;
+    });
+
+    it('should handle validation errors', async () => {
+      // Arrange
+      req.body = { fullName: '' }; // Empty name should trigger validation
+
+      // Act
+      await postEditClientName(req as RequestWithMiddleware, res as Response, next);
+
+      // Assert - Should configure form response with errors, not redirect
+      expect(redirectStub.called).to.be.false;
+      expect(renderStub.calledWith('case_details/edit-client-name.njk')).to.be.true;
+    });
+  });
+});
